extraction_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "extraction-service"
version = "1.0.0"
description = "API robusta para orquestar scraping y extracciones de TMS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0.post1",
    "pydantic==2.5.2",
    "pydantic-settings==2.1.0",
    "playwright==1.40.0",
    "google-cloud-storage==2.13.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8"
]
